Audit round 41: fifteen defects across the VR, DS3, audio, and macro arcs
Nine of these were adversarially verified against the reference sources
before any edit; one candidate was refuted that way and dropped.

Behavioral:
- Adaptive-trigger to impulse clamped a 0..65535 trigger to 255 instead
  of scaling it, so every pull past 0.4% travel read as zone 9 and every
  startPos gate passed.
- The Sony packers re-narrowed -(-32768) back to a short, packing full
  stick-down as full-up. Reachable through an AxisAdd macro that
  saturates an axis under a deflected stick.
- Persona haptics extracted into a static scratch from one pacing thread
  per persona, so two composite slots cross-fed each other's samples.
- The USB frame provider re-indexed the haptic ring dictionary after
  TryGetValue; a concurrent detach threw inside the WASAPI render thread
  and killed pad audio with nothing to notice it.
- The BT-mic and USB-jack readers reused one stop flag and one handle
  field across a stop-then-start with no join, so a retired reader could
  read and then close the live reader's handle. Both lanes are
  generation-scoped now and read their own handle.
- The VR mapping lane was missing from ComputeChecksum, CopyFrom,
  ToJson/FromJson, and CopyFromTranslated's flush: checksum dedup could
  collapse two devices onto one PadSetting, and profile saves and
  clipboard copies dropped VR mappings outright.
- The VR haptic handler checked _connected outside the lock Disconnect
  takes, so a pulse could re-latch a motor and build a timer nothing
  would dispose.
- MaxVrSlots was enforced only in the add popup. Every type-switch path
  bypassed it, and a second VR slot does not fail loudly: both consumers
  share one channel, latest writer wins. The cap now lives on the
  operation.
- BthPS3's RawPDO/ExclusivePDO were written unconditionally while only
  AutoEnableFilter honored the DsHidMini coexistence policy, which would
  kill their Bluetooth DS3s at the next re-enumeration.
- The DS3 suppress-reconnect flag was one bool for two overlapping
  flows, so whichever finished first un-suppressed the other.
- A DS3 that never answered its enable stayed attached and silent
  forever: five kicks, then a read loop that treats every pipe timeout
  as normal. It now detaches for a clean re-open. This is the reported
  "wired stops working after a pairing attempt until restart".
- GetFeature/SetFeature ignored the transferred count, so a short 0xF2
  reply yielded a zero MAC recorded as a successful pairing.
- SlotHasAnyMapping missed the KeepAwake legs its siblings gained.
- The Compass Yaw card gated on an axis-count flag the engine does not
  consume, offering the whole feature where it could never run.
- The rebuilt negative leg of a bipolar row dropped ParamAccel, so
  acceleration applied to one direction only.
- The headset gyro-live belief is revocable instead of a one-way latch.

Test isolation: StarterProfileRuntimeTests replaced the global settings
collections without joining SettingsManagerStatics.

Also in this commit: SteamVR installs Steam-free via steamcmd (first run
self-updates and exits without running its commands, so the invocation
is verify-gated and retried), the Dashboard driver card carries a
SteamVR row, and the seven-tile type ribbon fits again after the VR tile
pushed it past the card and rail widths.

+++ PadForge.App/Common/Input/SonyReportPackers.cs
@@ -97,9 +97,15 @@ namespace PadForge.Common.Input
             // Sticks (bytes 0-3): center 0x80. XInput Y is +up; DS4 firmware
             // is +down (HID convention) so Y axes are inverted.
             dest[0] = ToDs4Axis(gp.ThumbLX);
-            dest[1] = ToDs4Axis((short)-gp.ThumbLY);
+            // No (short) re-narrowing on the negation: -(-32768) is 32768,
+            // which wraps back to -32768 in a short and packs full-DOWN as
+            // 0x00 (full up). ToDs4Axis takes an int and clamps, so the
+            // widened value lands at 255 the way every other deflection
+            // does. Reachable whenever a source saturates an axis to
+            // short.MinValue (an AxisAdd macro on top of a deflected stick).
+            dest[1] = ToDs4Axis(-gp.ThumbLY);
             dest[2] = ToDs4Axis(gp.ThumbRX);
-            dest[3] = ToDs4Axis((short)-gp.ThumbRY);
+            dest[3] = ToDs4Axis(-gp.ThumbRY);
 
             // Buttons + hat (bytes 4-6).
             // byte 4: bits 0-3 = D-pad as 0..7 / 0x8=neutral; bits 4-7 = face buttons.
@@ -241,9 +247,15 @@ namespace PadForge.Common.Input
 
             // Sticks + triggers inline (bytes 0-5). Y inverted vs XInput.
             dest[0] = ToDs4Axis(gp.ThumbLX);
-            dest[1] = ToDs4Axis((short)-gp.ThumbLY);
+            // No (short) re-narrowing on the negation: -(-32768) is 32768,
+            // which wraps back to -32768 in a short and packs full-DOWN as
+            // 0x00 (full up). ToDs4Axis takes an int and clamps, so the
+            // widened value lands at 255 the way every other deflection
+            // does. Reachable whenever a source saturates an axis to
+            // short.MinValue (an AxisAdd macro on top of a deflected stick).
+            dest[1] = ToDs4Axis(-gp.ThumbLY);
             dest[2] = ToDs4Axis(gp.ThumbRX);
-            dest[3] = ToDs4Axis((short)-gp.ThumbRY);
+            dest[3] = ToDs4Axis(-gp.ThumbRY);
             dest[4] = (byte)(gp.LeftTrigger  >> 8);
             dest[5] = (byte)(gp.RightTrigger >> 8);
